requirements.txt
langchain==0.3.17
langchain-core==0.3.79
langchain-community==0.3.16
langchain-text-splitters==0.3.5

langchain-groq==0.2.1

faiss-cpu==1.7.4
sentence-transformers==2.7.0
torch==2.2.2

pydantic==2.7.4
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1

PyPDF2==3.0.1
jinja2==3.1.3
